+++ example-projects/java11/src/main/java/module-info.java
@@ -1,0 +1,4 @@
+module jgiven.exampleprojects.java11 {
+    exports com.tngtech.jgiven.exampleprojects.java11;
+    requires java.base;
+}