+++ jgiven-spock/src/test/groovy/com/tngtech/jgiven/spock/stages/WhenSomeAction.groovy
@@ -1,0 +1,10 @@
+package com.tngtech.jgiven.spock.stages
+
+import com.tngtech.jgiven.Stage
+
+class WhenSomeAction extends Stage<WhenSomeAction> {
+
+    WhenSomeAction some_action() {
+        self()
+    }
+}