+++ jgiven-spring/src/main/java/com/tngtech/jgiven/integration/spring/JGivenStage.java
@@ -6,10 +6,13 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
+import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 /**
- * Annotation that marks a bean as JGiven Stage.
+ * Annotation that marks a bean as JGiven Stage. This component will be in created in prototype scope
+ * to ensure that each test run receives a fresh (empty) instance of the bean.
  *
  * @since 0.8.0
  */
@@ -17,6 +20,7 @@ import org.springframework.stereotype.Component;
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 @Component
+@Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
 public @interface JGivenStage {
 
     /**